requirements.txt
aiogram<4.0
python-dotenv==1.0.0
yt-dlp==2024.8.6 
instaloader==4.13
TikTokApi==6.5.2